+++ web/test/helpers_test.dart
@@ -36,4 +36,23 @@ void main() {
     // Ensure accessing any arbitrary item in the list does not throw.
     expect(() => dartList[0], returnsNormally);
   });
+
+  test('responseHeaders transforms headers into a map', () async {
+    final request = XMLHttpRequest()
+      ..open('GET', 'www.google.com')
+      ..send();
+
+    await request.onLoad.first;
+
+    expect(
+      request.responseHeaders,
+      allOf(
+        containsPair('content-length', '10'),
+        containsPair('content-type', 'text/plain; charset=utf-8'),
+        containsPair('x-content-type-options', 'nosniff'),
+        containsPair('x-frame-options', 'SAMEORIGIN'),
+        containsPair('x-xss-protection', '1; mode=block'),
+      ),
+    );
+  });
 }